requirements.txt
flask
pandas
numpy
faiss-cpu
python-docx
openpyxl
openai
python-dotenv
